Dev2 (#28)
* Update specs.yaml

Fixed a typo in the specs.yaml file. The correct OpenAPI 2.0 format now applies for the /attractions endpoint.

* Added GCP env variables (#21)

Replaced URL placeholders with an env variable in the API surface document.

* Issue#20 aura db connection (#25)

* Added an Aura DB handler

* Issue#20 aura db connection (#22)

* Update specs.yaml

Fixed a typo in the specs.yaml file. The correct OpenAPI 2.0 format now applies for the /attractions endpoint.

* Added GCP env variables (#21)

Replaced URL placeholders with an env variable in the API surface document.

* Added an Aura DB handler

* Added a "Session" helper function

* Added db connection to main

Added a function to pass a Neo4j session to one of the GET methods.

* Issue#20 aura db connection (#26)

* Added an Aura DB handler

* Issue#20 aura db connection (#22)

* Update specs.yaml

Fixed a typo in the specs.yaml file. The correct OpenAPI 2.0 format now applies for the /attractions endpoint.

* Added GCP env variables (#21)

Replaced URL placeholders with an env variable in the API surface document.

* Added an Aura DB handler

* Added a "Session" helper function

* Added db connection to main

Added a function to pass a Neo4j session to one of the GET methods.

* Added a driver middleware function

* Issue#20 aura db connection (#27)

* Added an Aura DB handler

* Issue#20 aura db connection (#22)

* Update specs.yaml

Fixed a typo in the specs.yaml file. The correct OpenAPI 2.0 format now applies for the /attractions endpoint.

* Added GCP env variables (#21)

Replaced URL placeholders with an env variable in the API surface document.

* Added an Aura DB handler

* Added a "Session" helper function

* Added db connection to main

Added a function to pass a Neo4j session to one of the GET methods.

* Added a driver middleware function

* Added connection to AuraDB

Added a connection to AuraDB. The GET methods are now returning data from the database, instead of the test store.